refactor: Moved functions for code generation in to their own file
